uint256 >> uint

+++ contracts/fantasy.sol
@@ -9,23 +9,23 @@ import "@openzeppelin/contracts/token/ERC721/utils/ERC721Holder.sol";
 import "hardhat/console.sol";
 
 interface IPoints is IERC20 {
-    function mint(address to, uint256 amount) external;
+    function mint(address to, uint amount) external;
 }
 
 contract fantasy is Ownable, ERC721Holder {
     IPoints public points;
     IERC721 public player;
 
-    uint256 public stakedTotal;
-    uint256 public stakingStartTime;
-    uint256 constant stakingTime = 180 seconds;
-    uint256 constant token = 10e18;
+    uint public stakedTotal;
+    uint public stakingStartTime;
+    uint constant stakingTime = 180 seconds;
+    uint constant token = 10e18;
 
     struct Staker {
-        uint256[] tokenIds;
-        mapping(uint256 => uint256) tokenStakingCoolDown;
-        uint256 balance;
-        uint256 rewardsReleased;
+        uint[] tokenIds;
+        mapping(uint => uint) tokenStakingCoolDown;
+        uint balance;
+        uint rewardsReleased;
     }
 
     constructor(IERC721 _player, IPoints _points) {
@@ -37,26 +37,25 @@ contract fantasy is Ownable, ERC721Holder {
     mapping(address => Staker) public stakers;
 
     /// @notice Mapping from token ID to owner address
-    mapping(uint256 => address) public tokenOwner;
+    mapping(uint => address) public tokenOwner;
     bool public tokensClaimable;
     bool initialized;
 
     /// @notice event emitted when a user has staked a nft
 
-    event Staked(address owner, uint256 amount);
+    event Staked(address owner, uint amount);
 
     /// @notice event emitted when a user has unstaked a nft
-    event Unstaked(address owner, uint256 amount);
+    event Unstaked(address owner, uint amount);
 
     /// @notice event emitted when a user claims reward
-    event RewardPaid(address indexed user, uint256 reward);
+    event RewardPaid(address indexed user, uint reward);
 
     /// @notice Allows reward tokens to be claimed
     event ClaimableStatusUpdated(bool status);
 
     /// @notice Emergency unstake tokens without rewards
-    event EmergencyUnstake(address indexed user, uint256 tokenId);
-
+    event EmergencyUnstake(address indexed user, uint tokenId);
 
     function initStaking() public onlyOwner {
         //needs access control
@@ -76,18 +75,17 @@ contract fantasy is Ownable, ERC721Holder {
     function getScore(address _user)
         public
         view
-        returns (uint256[] memory tokenIds)
+        returns (uint[] memory tokenIds)
     {
         return stakers[_user].tokenIds;
     }
 
     // Setting the lineup
-    function stake(uint256 tokenId) public {
+    function stake(uint tokenId) public {
         _stake(msg.sender, tokenId);
     }
 
-
-    function _stake(address _user, uint256 _tokenId) internal {
+    function _stake(address _user, uint _tokenId) internal {
         require(initialized, "Fantasy Leauge: Football season has not started");
         require(
             player.ownerOf(_tokenId) == _user,
@@ -105,13 +103,13 @@ contract fantasy is Ownable, ERC721Holder {
         stakedTotal++;
     }
 
-    function unstake(uint256 _tokenId) public {
+    function unstake(uint _tokenId) public {
         claimReward(msg.sender);
         _unstake(msg.sender, _tokenId);
     }
 
     // Unstake without caring about rewards. EMERGENCY ONLY.
-    function emergencyUnstake(uint256 _tokenId) public {
+    function emergencyUnstake(uint _tokenId) public {
         require(
             tokenOwner[_tokenId] == msg.sender,
             "nft._unstake: Sender must have staked tokenID"
@@ -120,15 +118,15 @@ contract fantasy is Ownable, ERC721Holder {
         emit EmergencyUnstake(msg.sender, _tokenId);
     }
 
-    function _unstake(address _user, uint256 _tokenId) internal {
+    function _unstake(address _user, uint _tokenId) internal {
         require(
             tokenOwner[_tokenId] == _user,
             "Nft Staking System: user must be the owner of the staked nft"
         );
         Staker storage staker = stakers[_user];
 
-        //uint256 lastIndex = staker.tokenIds.length - 1;
-        //uint256 lastIndexKey = staker.tokenIds[lastIndex];
+        //uint lastIndex = staker.tokenIds.length - 1;
+        //uint lastIndexKey = staker.tokenIds[lastIndex];
 
         if (staker.tokenIds.length > 0) {
             staker.tokenIds.pop();
@@ -142,10 +140,9 @@ contract fantasy is Ownable, ERC721Holder {
         stakedTotal--;
     }
 
-    function updateReward(address _user, uint256 score) public {
-
+    function updateReward(address _user, uint score) public {
         Staker storage staker = stakers[_user];
-        staker.balance +=  token * score;
+        staker.balance += token * score;
         tokensClaimable == true;
 
         console.logUint(staker.balance);
@@ -154,8 +151,7 @@ contract fantasy is Ownable, ERC721Holder {
     // might remove this later
     function claimReward(address _user) public {
         require(tokensClaimable == true, "Tokens cannnot be claimed yet");
-        require(stakers[_user].balance > 0 , "0 rewards yet");
-
+        require(stakers[_user].balance > 0, "0 rewards yet");
 
         stakers[_user].rewardsReleased += stakers[_user].balance;
         stakers[_user].balance = 0;
@@ -164,5 +160,4 @@ contract fantasy is Ownable, ERC721Holder {
 
         emit RewardPaid(_user, stakers[_user].balance);
     }
-
 }
